% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonbgfa-package.R
\docType{package}
\name{nonbgfa-package}
\alias{nonbgfa}
\alias{nonbgfa-package}
\title{nonbgfa: Find Non-B DNA Forming Motifs in Genomic Sequences}
\description{
Identifies sequences associated with non-B DNA-forming motifs in FASTA-format genomic sequences. Detects seven motif classes: inverted repeats (cruciform DNA), mirror repeats (triplex DNA), direct repeats (slipped DNA), G-quadruplexes, Z-DNA, short tandem repeats (STR), and A-phased repeats (bent DNA). Based on the GFA suite developed at NCI-Frederick / Frederick National Laboratory for Cancer Research. See \url{https://nonb-abcc.ncifcrf.gov/apps/site/default}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abcsFrederick/non-b-gfa}
  \item Report bugs at \url{https://github.com/abcsFrederick/non-b-gfa/issues}
}

}
\author{
\strong{Maintainer}: Randy Johnson \email{johnson@hood.edu}

Authors:
\itemize{
  \item Duncan E. Donohue (Original GFA C implementation)
  \item Jack R. Collins (Original GFA C implementation)
}

}
\keyword{internal}
