% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_granges}
\alias{to_granges}
\title{Convert a nonbgfa data.frame to a GRanges object}
\usage{
to_granges(df)
}
\arguments{
\item{df}{A \code{data.frame} returned by one of the \verb{find_*()} functions.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object.
}
\description{
Requires the \pkg{GenomicRanges} package (Bioconductor).
}
